Split moving window analysis

Function SMW performs split moving window analysis (SMW) with randomizations tests. It may compute dissimilarities for a single window size or for several windows sizes.

SMW(yo, ws, dist = "bray", rand = c("shift", "plot"), n.rand = 99)

Arguments

yo
The ordered community matrix.

ws
The window sizes to be analyzed. Either a single value or a vector of values.

dist
The dissimilarity index used in vegan::vegdist. Defaults to 'bray'.

rand
The type of randomization for significance computation (Erdös et.al, 2014):

"shift": restricted randomization in which data belonging to the same species are randomly shifted along the data series ("Random shift");

"plot": unrestricted randomization: each sample is randomly repositioned along the data series ("Random plot").

n.rand
The number of randomizations.

Value

A two-level list object (class smw) describing the SMW results for each window w analyzed. The smw object is of length ws, and each of the w slots is a list of SMW results:

..$dp: The raw dissimilarity profile (DP). The DP is a data frame giving the positions, labels, values of dissimilarity and z-scores for each sample;

..$rdp: data frame containing the randomized DP;

..$md: mean dissimilarity of the randomized DP;

..$sd: standard deviation for each sample position;

..$oem: overall expected mean dissimilarity;

..$osd: average standard deviation for the dissimilarities;

..$params: list with input arguments

Available methods for class "smw" are print, extract and plot.

References

Erdos, L., Z. Bátori, C. S. Tölgyesi, and L. Körmöczi. 2014. The moving split window (MSW) analysis in vegetation science - An overview. Applied Ecology and Environmental Research 12:787–805.

Cornelius, J. M., and J. F. Reynolds. 1991. On Determining the Statistical Significance of Discontinuities with Ordered Ecological Data. Ecology 72:2057–2070.

See also

Examples

data(sim1)
sim1o<-OrdData(sim1$envi,sim1$comm)
ws50<-SMW(yo=sim1o$yo,ws=50, n=5)

#> 
#>  SMW analysis (1/1); w =50

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%

pool<-SMW(yo=sim1o$yo,ws=c(40,50), n.rand=5)

#> 
#>  SMW analysis (1/2); w =40

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%

#> 
#>  SMW analysis (2/2); w =50

#>   |                                                                              |                                                                      |   0%  |                                                                              |==============                                                        |  20%  |                                                                              |============================                                          |  40%  |                                                                              |==========================================                            |  60%  |                                                                              |========================================================              |  80%  |                                                                              |======================================================================| 100%