Extract results and breakpoints from a smw object

Functions to extract results and breakpoints from a smw object

# S3 method for smw
extract(smw, w = NULL, index = "dp", sig = "z",
  z = 1.85, BPs = "max", seq.sig = 3)

# S3 method for dp
bp(dp)

Arguments

smw
An object of class smw resulted from the SMW analysis.

w
Numeric. A "target" window size from which results will be extracted (see Details). Only effective if the smw object contains results from multiple window sizes.

index
The result to be extracted:

"dp": The dissimilarity profile (DP) table containing significant discontinuities and suggested breakpoints. The returned DP has class "dp", with an own generic print method. The print command prints and returns invisibly the DP.

"rdp": data frame containing the randomized DP;

"md": mean dissimilarity of the randomized DP;

"sd": standard deviation for each sample position;

"oem": overall expected mean dissimilarity;

"osd": average standard deviation for the dissimilarities;

"params": list with input arguments

sig
Significance test for detecting dissimilarity values that differs significantly from those appearing in a random pattern. The following tests are considered with default to sig.test="z":

"dp" consider normalized dissimilarity (z-scores) discontinuities that exceed a "z" critical value;

"sd" consider dissimilarity discontinuities that exceed mean plus one standard deviation;

"SD2" consider dissimilarity discontinuities that exceed mean plus two standard deviation;

"tail1" Consider dissimilarity discontinuities that exceed 95 percent confidence limits.

z
The critical value for the significance of z-values. Defaults to 'z=1.85' (Erdös et.al, 2014).

BPs
Defines if the breakpoints should be chosen as those sample positions corresponding to the maximum dissimilarity in a sequence of significant values ("max") or as those sample positions corresponding to the median position of the sequence ("median"). Defaults to BPs="max". If NULL the breakpoints are not computed.

seq.sig
The maximum length of consecutive, significant values of dissimilarity that will be considered in defining the community breakpoints. Defaults to seq.sig=3;

dp
An object of class dp (see Details).

Value

"extract" returns a result called by the argument index (see Details);

"bp" returns the locations of the breakpoints.

Details

If the smw object contains results from multiple window sizes, the DP table will be based on the average Z-score over the set of analysed window sizes. Available methods for class "dp" are print, bp and plot.

The argument w is optional. If the smw object is length 1, w is ignored. If length(smw)>1 and w is NULL, the function will extract the dissimilarity profile averaged over the set of window sizes.

See also

Examples

data(sim1)
sim1o<-OrdData(sim1$envi,sim1$comm)

ws50<-SMW(yo=sim1o$yo,ws=50)

#> 
#>  SMW analysis (1/1); w =50

#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%  |                                                                              |======                                                                |   9%  |                                                                              |=======                                                               |  10%  |                                                                              |========                                                              |  11%  |                                                                              |========                                                              |  12%  |                                                                              |=========                                                             |  13%  |                                                                              |==========                                                            |  14%  |                                                                              |===========                                                           |  15%  |                                                                              |===========                                                           |  16%  |                                                                              |============                                                          |  17%  |                                                                              |=============                                                         |  18%  |                                                                              |=============                                                         |  19%  |                                                                              |==============                                                        |  20%  |                                                                              |===============                                                       |  21%  |                                                                              |================                                                      |  22%  |                                                                              |================                                                      |  23%  |                                                                              |=================                                                     |  24%  |                                                                              |==================                                                    |  25%  |                                                                              |==================                                                    |  26%  |                                                                              |===================                                                   |  27%  |                                                                              |====================                                                  |  28%  |                                                                              |=====================                                                 |  29%  |                                                                              |=====================                                                 |  30%  |                                                                              |======================                                                |  31%  |                                                                              |=======================                                               |  32%  |                                                                              |=======================                                               |  33%  |                                                                              |========================                                              |  34%  |                                                                              |=========================                                             |  35%  |                                                                              |=========================                                             |  36%  |                                                                              |==========================                                            |  37%  |                                                                              |===========================                                           |  38%  |                                                                              |============================                                          |  39%  |                                                                              |============================                                          |  40%  |                                                                              |=============================                                         |  41%  |                                                                              |==============================                                        |  42%  |                                                                              |==============================                                        |  43%  |                                                                              |===============================                                       |  44%  |                                                                              |================================                                      |  45%  |                                                                              |=================================                                     |  46%  |                                                                              |=================================                                     |  47%  |                                                                              |==================================                                    |  48%  |                                                                              |===================================                                   |  49%  |                                                                              |===================================                                   |  51%  |                                                                              |====================================                                  |  52%  |                                                                              |=====================================                                 |  53%  |                                                                              |=====================================                                 |  54%  |                                                                              |======================================                                |  55%  |                                                                              |=======================================                               |  56%  |                                                                              |========================================                              |  57%  |                                                                              |========================================                              |  58%  |                                                                              |=========================================                             |  59%  |                                                                              |==========================================                            |  60%  |                                                                              |==========================================                            |  61%  |                                                                              |===========================================                           |  62%  |                                                                              |============================================                          |  63%  |                                                                              |=============================================                         |  64%  |                                                                              |=============================================                         |  65%  |                                                                              |==============================================                        |  66%  |                                                                              |===============================================                       |  67%  |                                                                              |===============================================                       |  68%  |                                                                              |================================================                      |  69%  |                                                                              |=================================================                     |  70%  |                                                                              |=================================================                     |  71%  |                                                                              |==================================================                    |  72%  |                                                                              |===================================================                   |  73%  |                                                                              |====================================================                  |  74%  |                                                                              |====================================================                  |  75%  |                                                                              |=====================================================                 |  76%  |                                                                              |======================================================                |  77%  |                                                                              |======================================================                |  78%  |                                                                              |=======================================================               |  79%  |                                                                              |========================================================              |  80%  |                                                                              |=========================================================             |  81%  |                                                                              |=========================================================             |  82%  |                                                                              |==========================================================            |  83%  |                                                                              |===========================================================           |  84%  |                                                                              |===========================================================           |  85%  |                                                                              |============================================================          |  86%  |                                                                              |=============================================================         |  87%  |                                                                              |==============================================================        |  88%  |                                                                              |==============================================================        |  89%  |                                                                              |===============================================================       |  90%  |                                                                              |================================================================      |  91%  |                                                                              |================================================================      |  92%  |                                                                              |=================================================================     |  93%  |                                                                              |==================================================================    |  94%  |                                                                              |==================================================================    |  95%  |                                                                              |===================================================================   |  96%  |                                                                              |====================================================================  |  97%  |                                                                              |===================================================================== |  98%  |                                                                              |===================================================================== |  99%  |                                                                              |======================================================================| 100%

ws50_dp<-extract(ws50)

#> Selected index: 'dp'

#> 1 window size(s) available in 'smw': 50

#> Window size selected: 50

#> Number of breakpoints detected: 1

#> Breakpoint positions:  51

head(ws50_dp)

#>   positions sampleID      diss   zscore sig:z bp
#> 1        25       41 0.1530606 3.380557     *  -
#> 2        26       40 0.1645366 4.075955     *  -
#> 3        27       98 0.1755934 4.745946     *  -
#> 4        28        7 0.1854114 5.340874     *  -
#> 5        29       61 0.1983146 6.122756     *  -
#> 6        30       51 0.2165018 7.224820     *  -